Add MainScreen component that deals with switching the type of editor
